ACC-3028 move ContentStoreResolver to contentgrid-appserver-contentstore-api
