+++ contentgrid-appserver-autoconfigure/src/test/java/com/contentgrid/appserver/autoconfigure/contentstore/S3ContentStoreAutoConfigurationTest.java
@@ -2,7 +2,7 @@ package com.contentgrid.appserver.autoconfigure.contentstore;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.contentgrid.appserver.domain.spi.contentstore.resolver.ContentStoreResolver;
+import com.contentgrid.appserver.contentstore.api.resolver.ContentStoreResolver;
 import io.minio.MinioAsyncClient;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
