+++ contentgrid-appserver-contentstore-api/build.gradle
@@ -5,5 +5,6 @@ plugins {
 }
 
 dependencies {
+    api project(':contentgrid-appserver-application-model')
 }
