+++ contentgrid-appserver-integration-test/src/test/java/com/contentgrid/appserver/integration/test/content/EncryptedAlfCompatibilityTest.java
@@ -13,12 +13,12 @@ import com.contentgrid.appserver.application.model.values.LinkName;
 import com.contentgrid.appserver.application.model.values.PathSegmentName;
 import com.contentgrid.appserver.application.model.values.TableName;
 import com.contentgrid.appserver.contentstore.api.ContentAccessor;
+import com.contentgrid.appserver.contentstore.api.resolver.ContentStoreResolver;
 import com.contentgrid.appserver.contentstore.impl.encryption.engine.DataEncryptionAlgorithm;
 import com.contentgrid.appserver.contentstore.impl.encryption.keys.DataEncryptionKeyAccessor;
 import com.contentgrid.appserver.contentstore.impl.encryption.keys.KeyBytes;
 import com.contentgrid.appserver.contentstore.impl.encryption.keys.StoredDataEncryptionKey;
 import com.contentgrid.appserver.contentstore.impl.encryption.keys.WrappingKeyId;
-import com.contentgrid.appserver.domain.spi.contentstore.resolver.ContentStoreResolver;
 import com.contentgrid.appserver.query.engine.api.CreateEventConsumer;
 import com.contentgrid.appserver.query.engine.api.DeleteEventConsumer;
 import com.contentgrid.appserver.query.engine.api.LinkEventConsumer;
