+++ contentgrid-appserver-rest/build.gradle
@@ -18,7 +18,6 @@ dependencies {
     implementation 'org.springframework.security:spring-security-oauth2-resource-server'
 
     testImplementation project(':contentgrid-appserver-spring-boot-starter')
-    testImplementation project(':contentgrid-appserver-domain-spi')
     testImplementation testFixtures(project(':contentgrid-appserver-contentstore-impl-utils'))
     testImplementation testFixtures(project(':contentgrid-appserver-application-model'))
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
