+++ contentgrid-appserver-rest/src/test/java/com/contentgrid/appserver/rest/entity/ContentRestControllerTest.java
@@ -16,8 +16,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.contentgrid.appserver.contentstore.api.ContentReference;
 import com.contentgrid.appserver.contentstore.api.ContentStore;
 import com.contentgrid.appserver.contentstore.api.UnwritableContentException;
+import com.contentgrid.appserver.contentstore.api.resolver.ContentStoreResolver;
 import com.contentgrid.appserver.contentstore.impl.utils.testing.MockContentStore;
-import com.contentgrid.appserver.domain.spi.contentstore.resolver.ContentStoreResolver;
 import com.contentgrid.appserver.rest.test.TestApplication;
 import com.contentgrid.appserver.query.engine.api.TableCreator;
 import com.contentgrid.appserver.rest.test.ProblemDetailsMockMvcMatchers;
